Implement password management and configuration initialization in ConfigManager
- Added initialization of default configuration settings in ConfigManager, including sound settings and password management.
- Refactored password prompt handling to be encapsulated within the ConfigManager class, improving modularity.
- Removed redundant password prompt logic from main.js, streamlining the application flow.
- Updated sound configuration retrieval to align with the new structure in ConfigManager.
