+++ configManager.js
@@ -1,10 +1,34 @@
 const fs = require('node:fs');
 const path = require('node:path');
+const { BrowserWindow, ipcMain } = require('electron');
 
 class ConfigManager {
     constructor() {
         this.configPath = path.join(__dirname, "wg_config.json");
         this.config = this.loadConfig();
+        this.initializeConfig();
+    }
+
+    initializeConfig() {
+        const defaultConfig = {
+            sound: {
+                enabled: true
+            },
+            password: null
+        };
+
+        // Merge default config with existing config, preserving any existing values
+        this.config = {
+            ...defaultConfig,
+            ...this.config,
+            sound: {
+                ...defaultConfig.sound,
+                ...(this.config.sound || {})
+            }
+        };
+
+        // Save the merged config
+        this.saveConfig();
     }
 
     loadConfig() {
@@ -39,11 +63,12 @@ class ConfigManager {
 
     // Sound configuration
     getSoundEnabled() {
-        return this.config.soundEnabled !== false;
+        return this.config.sound.enabled;
     }
 
     setSoundEnabled(enabled) {
-        this.set('soundEnabled', enabled);
+        this.config.sound.enabled = enabled;
+        this.saveConfig();
     }
 
     // Password configuration
@@ -52,8 +77,87 @@ class ConfigManager {
     }
 
     setPassword(password) {
-        this.set('password', password);
+        this.config.password = password;
+        this.saveConfig();
+    }
+
+    async checkPasswordConfig() {
+        if (!this.config.password && this.config.password !== "") {
+            const password = await this.promptForPassword();
+            this.setPassword(password || "");
+        }
+    }
+
+    promptForPassword() {
+        return new Promise((resolve, _reject) => {
+            const channelId = `password-prompt-response-${Date.now()}-${Math.floor(Math.random() * 1000)}`;
+            const promptWindow = new BrowserWindow({
+                width: 300,
+                height: 250,
+                title: "Configure Viewer Password",
+                parent: require('./windowManager').getMainWindow(),
+                modal: true,
+                show: false,
+                webPreferences: {
+                    preload: path.join(__dirname, 'promptPreload.js'),
+                    nodeIntegration: false,
+                    contextIsolation: true,
+                    webSecurity: true
+                }
+            });
+
+            // Read the CSS file and inline its contents
+            const styleContent = fs.readFileSync(path.join(__dirname, 'styles.css'), 'utf-8');
+
+            const htmlContent = `<!DOCTYPE html>
+<html>
+    <head>
+        <meta name="response-channel" content="${channelId}">
+        <title>Configure Viewer Password</title>
+        <style>${styleContent}</style>
+    </head>
+    <body>
+        <p class="prompt-message">Enter a password for the viewer window. Leave blank for no password:</p>
+        <input id="pwd" type="password" autofocus />
+        <div class="button-container">
+            <button id="submit">Submit</button>
+            <button id="cancel">Cancel</button>
+        </div>
+        <script>
+            const responseChannel = "${channelId}";
+            document.getElementById('submit').addEventListener('click', () => {
+                const value = document.getElementById('pwd').value;
+                window.Electron.sendResponse(responseChannel, value);
+            });
+            document.getElementById('cancel').addEventListener('click', () => {
+                window.Electron.sendResponse(responseChannel, null);
+            });
+            document.addEventListener('keydown', (event) => {
+                if (event.key === 'Enter') {
+                    event.preventDefault();
+                    const value = document.getElementById('pwd').value;
+                    window.Electron.sendResponse(responseChannel, value);
+                } else if (event.key === 'Escape') {
+                    event.preventDefault();
+                    window.Electron.sendResponse(responseChannel, null);
+                }
+            });
+        </script>
+    </body>
+</html>`;
+
+            promptWindow.loadURL(`data:text/html;charset=utf-8,${encodeURIComponent(htmlContent)}`);
+            promptWindow.once('ready-to-show', () => {
+                promptWindow.show();
+            });
+            ipcMain.once(channelId, (event, value) => {
+                resolve(value);
+                if (!promptWindow.isDestroyed()) {
+                    promptWindow.close();
+                }
+            });
+        });
     }
 }
 
-module.exports = new ConfigManager(); 
+module.exports = new ConfigManager();
